Added auth to teamService, changed auth in ocelot.json, fixed auth0 in frontend

+++ src/backend/Playprism/Services/TeamService/Playprism.Services.TeamService.API/Controllers/TeamController.cs
@@ -1,29 +1,33 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Playprism.Services.TeamService.API.Entities;
 using Playprism.Services.TeamService.API.Interfaces.Services;
-using System;
 using System.Collections.Generic;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Playprism.Services.TeamService.API.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/v1/[controller]")]
     public class TeamController : ControllerBase
     {
         private readonly ITeamManageService _teamManageService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public TeamController(ITeamManageService teamManageService)
+        public TeamController(ITeamManageService teamManageService, IAuthorizationService authorizationService)
         {
             _teamManageService = teamManageService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
         [ProducesResponseType(typeof(IEnumerable<TeamEntity>), StatusCodes.Status200OK)]
         public async Task<ActionResult<IEnumerable<TeamEntity>>> Get()
         {
-            int userId = new Random().Next(10); // todo: get userId from token
+            string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             var entities = await _teamManageService.GetTeamsAsync(userId);
 
             return Ok(entities);
@@ -47,6 +51,7 @@ namespace Playprism.Services.TeamService.API.Controllers
         [ProducesResponseType(typeof(TeamEntity), StatusCodes.Status202Accepted)]
         public async Task<ActionResult<TeamEntity>> Post([FromBody] TeamEntity entity)
         {
+            entity.OwnerId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             entity = await _teamManageService.AddTeamAsync(entity);
 
             return Accepted(entity);
@@ -54,8 +59,20 @@ namespace Playprism.Services.TeamService.API.Controllers
 
         [HttpPut("{id}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult> Put(int id, [FromBody] TeamEntity entity)
         {
+            var team = await _teamManageService.GetTeamAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, team, "TeamOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+            if (team == null)
+            {
+                return NotFound();
+            }
+
             await _teamManageService.UpdateTeamAsync(entity);
 
             return NoContent();
@@ -63,13 +80,20 @@ namespace Playprism.Services.TeamService.API.Controllers
 
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status202Accepted)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult> Delete(int id)
         {
             var team = await _teamManageService.GetTeamAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, team, "TeamOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
             if (team == null)
             {
                 return NotFound();
             }
+
             await _teamManageService.DeleteTeamAsync(team);
 
             return Accepted();
@@ -79,17 +103,26 @@ namespace Playprism.Services.TeamService.API.Controllers
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<ActionResult> InvitePlayerToTeam(int id, string username)
         {
+            var team = await _teamManageService.GetTeamAsync(id);
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, team, "TeamOwnerPolicy");
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+            if (team == null)
+            {
+                return NotFound();
+            }
             await _teamManageService.InvitePlayerAsync(id, username);
 
             return Accepted();
         }
 
-
         [HttpPost("{id}/join")]
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<ActionResult<TeamEntity>> JoinTeam(int id)
         {
-            int userId = new Random().Next(10); // todo: get userId from token
+            string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             await _teamManageService.JoinTeamAsync(userId, id);
 
             return Accepted();
@@ -99,7 +132,7 @@ namespace Playprism.Services.TeamService.API.Controllers
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<ActionResult> RefuseTeam(int id)
         {
-            int userId = new Random().Next(10); // todo: get userId from token
+            string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             await _teamManageService.JoinTeamAsync(userId, id);
 
             return Accepted();
@@ -109,7 +142,7 @@ namespace Playprism.Services.TeamService.API.Controllers
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         public async Task<ActionResult<TeamEntity>> LeaveTeam(int id)
         {
-            int userId = new Random().Next(10); // todo: get userId from token
+            string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
             await _teamManageService.LeaveTeamAsync(userId, id);
 
             return Accepted();
